Rename new implementations, remove old ones
