+++ toon/input/__init__.py
@@ -1,2 +1,3 @@
 from .hand import Hand
+from .birds import BlamBirds
 from .base_input import BaseInput, DummyTime
